panda daemon: fix
